Add `dired-mark` face.
Default `dired-mark` face is inherited from `error`, which color is orange, the same as `dired-directory` which is inherited from `font-lock-function-name-face`.

https://github.com/purcell/color-theme-sanityinc-tomorrow/blob/master/color-theme-sanityinc-tomorrow.el#L186

https://github.com/purcell/color-theme-sanityinc-tomorrow/blob/master/color-theme-sanityinc-tomorrow.el#L200

It's hard for me to distinguish whether I had marked or not, so I add a face to it. I noticed there is a selection face, so I use it here.